fix(frontend): resolve type errors blocking the tsc gate
Fixes every error tsc --noEmit and Next.js page typing report, ahead
of re-arming the TypeScript build gate:

- contact form: split zod input/output types (is_primary default) and
  type onSubmit with the transformed payload, not raw form data
- time-acceleration widget: variant='default' matched no Button
  variant, so active preset buttons rendered unstyled; use 'primary'
  (locked by a new unit test)
- settings/mac page: move renderCursorCell and RotateKeyModal to their
  own modules — pages may not have extra exports
- test-api page: type the overdue response envelope
- error-boundary test: vi.stubEnv instead of assigning readonly
  process.env.NODE_ENV
- method-conflict test: drop fixture fields that don't exist on
  ContactMethod (verified against the backend serializer)
- contact-tasks-api: spread interface-typed params into a plain object
- test-map: map changed frontend sources + unit-test dirs
