+++ frontend/src/app/settings/mac/__tests__/rotate-key-modal.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, fireEvent, waitFor } from '@testing-library/react'
 import { describe, expect, it, vi } from 'vitest'
 
-import { RotateKeyModal } from '../page'
+import { RotateKeyModal } from '../rotate-key-modal'
 
 describe('RotateKeyModal', () => {
   const token = {
